refactor(Landing Page): re - organised repo
