+++ src/app/[locale]/(auth)/_components/FooterWebApp.tsx
