+++ src/app/[locale]/(auth)/_components/Header.tsx
