+++ src/app/[locale]/(auth)/explore-story/[id]/_components/RatingOverview.tsx
