+++ src/app/[locale]/(auth)/home/_components/Banner.tsx
