+++ src/app/[locale]/(unauth)/(landingpage)/_components/About.tsx
