+++ src/app/[locale]/(unauth)/(landingpage)/_components/FAQs.tsx
