+++ src/app/[locale]/(unauth)/(landingpage)/_components/Footer.tsx
