+++ src/app/[locale]/(unauth)/(landingpage)/_components/HumanBookBanner.tsx
