+++ src/app/[locale]/(unauth)/(landingpage)/_components/News.tsx
