+++ src/app/[locale]/(unauth)/(landingpage)/_components/Newsletter.tsx
