+++ src/app/[locale]/(unauth)/(landingpage)/_components/OurMascot.tsx
