+++ src/app/[locale]/(unauth)/(landingpage)/_components/Sponsors.tsx
