+++ src/app/[locale]/(unauth)/(landingpage)/_components/hero/HeroDesktop.tsx
