+++ src/app/[locale]/(unauth)/(landingpage)/_components/hero/HeroMobile.tsx
