+++ src/app/[locale]/(unauth)/(landingpage)/_components/hero/index.tsx
