+++ src/features/stories/components/DetailedStory.tsx
