+++ src/features/stories/components/DetailedStoryWithCover.tsx
